fix: prevent name change form submission when browser tab is closed
When the browser tab is closed, a blur event occurs, and the name change
form may be submitted. If this happens, the username is saved
in local storage even though the username change has been completed.
To improve this, modify the code so that the form is not submitted
when the browser tab is closed.

Refs: #279