fix: fix issue where bio cannot be expanded when changing to long text
When changing the bio from a short text that does not need
to be collapsed to a long text that needs to be collapsed,
the issue where the bio cannot be expanded occurred.
This is because the rendering of BioCollapsibleSection is completed
before the revalidation by revalidatePath is finished.
Therefore, I fix it to pass the bio to the key props
so that BioCollapsibleSection is re-rendered after the revalidation
of the bio is completed.

Refs: #285